+++ TOWER_TECH_NOTES.md
@@ -42,82 +42,25 @@ const uint8_t lookup_led_channel[] = {
 
 ## Constants Reference
 
-```typescript
-const TOWER_LAYERS = {
-    TOP_RING: 0,
-    MIDDLE_RING: 1,
-    BOTTOM_RING: 2,
-    LEDGE: 3,
-    BASE1: 4,
-    BASE2: 5,
-} as const;
+Tower layer and position constants are defined in `src/udtConstants.ts`:
 
-// Ring layers use cardinal directions (position 0 = North)
-const RING_LIGHT_POSITIONS = {
-    NORTH: 0,
-    EAST: 1,
-    SOUTH: 2,
-    WEST: 3,
-} as const;
+- `TOWER_LAYERS` - Maps layer names to indices (0-5)
+- `RING_LIGHT_POSITIONS` - Cardinal direction positions for ring layers
+- `LEDGE_BASE_LIGHT_POSITIONS` - Ordinal direction positions for ledge/base layers  
+- `LED_CHANNEL_LOOKUP` - Hardware LED channel mapping array
+- `STATE_DATA_LENGTH` - Binary state data length (19 bytes)
 
-// Ledge and Base layers use ordinal directions (position 0 = North-East)
-const LEDGE_BASE_LIGHT_POSITIONS = {
-    NORTH_EAST: 0,
-    SOUTH_EAST: 1,
-    SOUTH_WEST: 2,
-    NORTH_WEST: 3,
-} as const;
-
-// LED Channel Lookup (matches firmware implementation)
-const LED_CHANNEL_LOOKUP = [
-    // Layer 0: Top Ring (C0 R0, C0 R3, C0 R2, C0 R1)
-    0, 3, 2, 1,
-    // Layer 1: Middle Ring (C1 R3, C1 R2, C1 R1, C1 R0)
-    7, 6, 5, 4,
-    // Layer 2: Bottom Ring (C2 R2, C2 R1, C2 R0, C2 R3)
-    10, 9, 8, 11,
-    // Layer 3: Ledge (LEDGE R4, LEDGE R5, LEDGE R6, LEDGE R7)
-    12, 13, 14, 15,
-    // Layer 4: Base1 (BASE1 R4, BASE1 R5, BASE1 R6, BASE1 R7)
-    16, 17, 18, 19,
-    // Layer 5: Base2 (BASE2 R4, BASE2 R5, BASE2 R6, BASE2 R7)
-    20, 21, 22, 23,
-];
-
-// Corrected mapping function
-function getTowerPosition(layerIndex: number, lightIndex: number) {
-    const isRingLayer = layerIndex <= 2;
-
-    if (isRingLayer) {
-        const directions = ['NORTH', 'EAST', 'SOUTH', 'WEST'];
-        const layerNames = ['TOP_RING', 'MIDDLE_RING', 'BOTTOM_RING'];
-        return {
-            level: layerNames[layerIndex],
-            direction: directions[lightIndex],
-            ledChannel: LED_CHANNEL_LOOKUP[layerIndex * 4 + lightIndex],
-        };
-    } else {
-        const directions = ['NORTH_EAST', 'SOUTH_EAST', 'SOUTH_WEST', 'NORTH_WEST'];
-        const layerNames = ['LEDGE', 'BASE1', 'BASE2'];
-        return {
-            level: layerNames[layerIndex - 3],
-            direction: directions[lightIndex],
-            ledChannel: LED_CHANNEL_LOOKUP[layerIndex * 4 + lightIndex],
-        };
-    }
-}
-```
+See the source file for the complete definitions and latest values.
 
 ## Usage Example
 
 ```typescript
+import { getTowerPosition, getActiveLights } from './udtHelpers';
 import {
-    getTowerPosition,
-    getActiveLights,
     TOWER_LAYERS,
     RING_LIGHT_POSITIONS,
     LEDGE_BASE_LIGHT_POSITIONS,
-} from './functions';
+} from './udtConstants';
 
 // Get position for a ring light
 const topNorth = getTowerPosition(TOWER_LAYERS.TOP_RING, RING_LIGHT_POSITIONS.NORTH);
@@ -155,7 +98,7 @@ The tower communication uses **20-byte command packets** that consist of:
 -   **Byte 0**: Command Type (always `0x00` for tower state commands)
 -   **Bytes 1-19**: Tower State Data (19 bytes containing complete tower state)
 
-The 19-byte tower state data is packed/unpacked using the `rtdt_pack_state` and `rtdt_unpack_state` functions from `functions.ts`, which match the firmware implementation exactly.
+The 19-byte tower state data is packed/unpacked using the `rtdt_pack_state` and `rtdt_unpack_state` functions from `udtTowerState.ts`, which match the firmware implementation exactly.
 
 ## Complete 20-Byte Command Packet Structure
